invoice-API/requirements.txt
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pika==1.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
reportlab==4.4.2
setuptools==80.9.0
Werkzeug==3.1.3
